Moved dataset preprocessors to a utils file instead of specific to VGG16
