+++ training/utils.py
@@ -1,11 +1,13 @@
 import math
 import os
 from datetime import datetime
-from typing import List, Optional, Tuple
+from typing import Callable, Iterable, List, Optional, Tuple
 
 import matplotlib.pyplot as plt
+import numpy as np
 import tensorflow as tf
 
+from data.dataset_utils import get_num_classes, pass_class_names
 from training import MODELS_DIR_ABS_PATH
 
 
@@ -53,3 +55,65 @@ def plot_batch_predictions(
         )
     fig.tight_layout()  # Prevent title overlap
     plt.show()
+
+
+def make_preprocessing_generator(
+    dataset: tf.data.Dataset,
+    num_repeat: int = -1,
+    image_preprocessor: Optional[Callable[[tf.Tensor], tf.Tensor]] = None,
+) -> Iterable[Tuple[tf.Tensor, np.ndarray]]:
+    """
+    Make an iterator that pre-processes the input dataset for training.
+
+    Args:
+        dataset: TensorFlow dataset to preprocess.
+        num_repeat: Optional count of dataset repetitions.
+            Default of -1 will repeat indefinitely.
+            For training data: set to the number of epochs, or -1.
+        image_preprocessor: Function to pre-process images per a model's requirements.
+            Default of None will not try to pre-process images.
+            Examples:
+            - VGG16: tf.keras.applications.vgg16.preprocess_input.
+            - ResNet50: tf.keras.applications.resnet50.preprocess_input.
+
+    Returns:
+        Iterable of (image, categorical label) tuples.
+    """
+    num_classes = get_num_classes(dataset)
+    for batch_images, batch_labels in dataset.repeat(num_repeat):
+        if image_preprocessor is not None:
+            batch_images = image_preprocessor(batch_images)
+        yield batch_images, tf.keras.utils.to_categorical(
+            batch_labels, num_classes, dtype="bool"
+        )
+
+
+def preprocess_dataset(
+    dataset: tf.data.Dataset,
+    image_preprocessor: Optional[Callable[[tf.Tensor], tf.Tensor]] = None,
+) -> tf.data.Dataset:
+    """
+    Preprocess the input dataset for training.
+
+    Args:
+        dataset: TensorFlow dataset to preprocess.
+        image_preprocessor: Function to pre-process images per a model's requirements.
+            Default of None will not try to pre-process images.
+            Examples:
+            - VGG16: tf.keras.applications.vgg16.preprocess_input.
+            - ResNet50: tf.keras.applications.resnet50.preprocess_input.
+
+    Returns:
+        Preprocessed dataset.
+    """
+    num_classes = get_num_classes(dataset)
+
+    def _preprocess(x: tf.Tensor, y: tf.Tensor) -> Tuple[tf.Tensor, tf.Tensor]:
+        if image_preprocessor is not None:
+            x = image_preprocessor(x)
+        # NOTE: one_hot is a transformation step for Tensors, so we use it here
+        # over to_categorical
+        # pylint: disable=no-value-for-parameter
+        return x, tf.one_hot(y, depth=num_classes)
+
+    return pass_class_names(dataset, dataset.map(_preprocess))
